Split auditLog types
