+++ packages/types/src/discord.ts
@@ -1,16 +1,9 @@
 import type { DiscordApplication, DiscordApplicationIntegrationType } from './discord/applications.js'
 import type { DiscordPresenceUpdate } from './discord/gateway.js'
-import type {
-  DiscordApplicationCommand,
-  DiscordApplicationCommandPermissions,
-  DiscordMessageComponents,
-  DiscordMessageInteraction,
-  InteractionTypes,
-} from './discord/interactions.js'
+import type { DiscordMessageComponents, DiscordMessageInteraction, InteractionTypes } from './discord/interactions.js'
 import type {
   AllowedMentionsTypes,
   AttachmentFlags,
-  AuditLogEvents,
   ChannelFlags,
   ChannelTypes,
   DefaultMessageNotificationLevels,
@@ -1472,29 +1465,6 @@ export interface DiscordListArchivedThreads extends DiscordListActiveThreads {
   has_more: boolean
 }
 
-/** https://discord.com/developers/docs/resources/audit-log#audit-log-object */
-export interface DiscordAuditLog {
-  /** List of webhooks found in the audit log */
-  webhooks: DiscordWebhook[]
-  /** List of users found in the audit log */
-  users: DiscordUser[]
-  /** List of audit log entries, sorted from most to least recent */
-  audit_log_entries: DiscordAuditLogEntry[]
-  /** List of partial integration objects */
-  integrations: Partial<DiscordIntegration>[]
-  /**
-   * List of threads found in the audit log.
-   * Threads referenced in `THREAD_CREATE` and `THREAD_UPDATE` events are included in the threads map since archived threads might not be kept in memory by clients.
-   */
-  threads: DiscordChannel[]
-  /** List of guild scheduled events found in the audit log */
-  guild_scheduled_events?: DiscordScheduledEvent[]
-  /** List of auto moderation rules referenced in the audit log */
-  auto_moderation_rules?: DiscordAutoModerationRule[]
-  /** List of application commands referenced in the audit log */
-  application_commands: DiscordApplicationCommand[]
-}
-
 /** https://discord.com/developers/docs/resources/auto-moderation#auto-moderation-rule-object */
 export interface DiscordAutoModerationRule {
   /** The id of this rule */
@@ -1636,164 +1606,6 @@ export interface DiscordAutoModerationActionMetadata {
   custom_message?: string
   /** Timeout duration in seconds maximum of 2419200 seconds (4 weeks). Only supported for TriggerType.Keyword && Only in ActionType.Timeout */
   duration_seconds?: number
-}
-
-/** https://discord.com/developers/docs/resources/audit-log#audit-log-entry-object-audit-log-entry-structure */
-export interface DiscordAuditLogEntry {
-  /** ID of the affected entity (webhook, user, role, etc.) */
-  target_id: string | null
-  /** Changes made to the `target_id` */
-  changes?: DiscordAuditLogChange[]
-  /** User or app that made the changes */
-  user_id: string | null
-  /** ID of the entry */
-  id: string
-  /** Type of action that occurred */
-  action_type: AuditLogEvents
-  /** Additional info for certain event types */
-  options?: DiscordOptionalAuditEntryInfo
-  /** Reason for the change (1-512 characters) */
-  reason?: string
-}
-
-export type DiscordAuditLogChangeObject<T> = {
-  [K in keyof T]: {
-    new_value?: T[K]
-    old_value?: T[K]
-    key: K
-  }
-}[keyof T]
-
-// Done manually as it is clearer in this way
-/** Partial role audit log entry change exception */
-export type DiscordAuditLogChangePartialRole = { new_value: { id: string; name: string }[]; key: '$add' | '$remove' }
-
-/**
- * Invite audit log entry change exception
- *
- * @remarks
- * While the docs say that 'channel.id' will never exist, we keep it as it is very complex to remove, the user should not use it and use the provided 'channel_id' value instead
- */
-export type DiscordAuditLogChangeInvite = DiscordAuditLogChangeObject<DiscordInvite> | DiscordAuditLogChangeObject<{ channel_id: string }>
-
-/**
- * Invite Metadata audit log entry change exception
- *
- * @remarks
- * While the docs say that 'channel.id' will never exist, we keep it as it is very complex to remove, the user should not use it and use the provided 'channel_id' value instead
- */
-export type DiscordAuditLogChangeInviteMetadata =
-  | DiscordAuditLogChangeObject<DiscordInviteMetadata>
-  | DiscordAuditLogChangeObject<{ channel_id: string }>
-
-/** Webhook audit log entry change exception */
-export type DiscordAuditLogChangeWebhook =
-  | DiscordAuditLogChangeObject<Omit<DiscordWebhook, 'avatar'>>
-  | DiscordAuditLogChangeObject<{ avatar_hash: DiscordWebhook['avatar'] }>
-
-// Done manually as it is clearer in this way
-/** Command Permission audit log entry change exception */
-export type DiscordAuditLogChangeApplicationCommandPermissions = {
-  new_value?: DiscordApplicationCommandPermissions
-  old_value?: DiscordApplicationCommandPermissions
-  key: `${number}`
-}
-
-/** https://discord.com/developers/docs/resources/audit-log#audit-log-change-object-audit-log-change-structure */
-export type DiscordAuditLogChange =
-  | DiscordAuditLogChangeObject<DiscordGuild>
-  | DiscordAuditLogChangeObject<DiscordChannel>
-  | DiscordAuditLogChangeObject<DiscordOverwrite>
-  | DiscordAuditLogChangeObject<DiscordMember>
-  | DiscordAuditLogChangePartialRole
-  | DiscordAuditLogChangeObject<DiscordRole>
-  | DiscordAuditLogChangeInvite
-  | DiscordAuditLogChangeInviteMetadata
-  | DiscordAuditLogChangeWebhook
-  | DiscordAuditLogChangeObject<DiscordEmoji>
-  | DiscordAuditLogChangeObject<DiscordIntegration>
-  | DiscordAuditLogChangeObject<DiscordStageInstance>
-  | DiscordAuditLogChangeObject<DiscordSticker>
-  | DiscordAuditLogChangeObject<DiscordThreadMetadata>
-  | DiscordAuditLogChangeApplicationCommandPermissions
-  | DiscordAuditLogChangeObject<DiscordAutoModerationRule>
-  | DiscordAuditLogChangeObject<DiscordGuildOnboardingPrompt>
-  | DiscordAuditLogChangeObject<DiscordGuildOnboarding>
-
-/** https://discord.com/developers/docs/resources/audit-log#audit-log-entry-object-optional-audit-entry-info */
-export interface DiscordOptionalAuditEntryInfo {
-  /**
-   * ID of the app whose permissions were targeted.
-   *
-   * Event types: `APPLICATION_COMMAND_PERMISSION_UPDATE`
-   */
-  application_id?: string
-  /**
-   * Name of the Auto Moderation rule that was triggered.
-   *
-   * Event types: `AUTO_MODERATION_BLOCK_MESSAGE`, `AUTO_MODERATION_FLAG_TO_CHANNEL`, `AUTO_MODERATION_USER_COMMUNICATION_DISABLED`
-   */
-  auto_moderation_rule_name?: string
-  /**
-   * Trigger type of the Auto Moderation rule that was triggered.
-   *
-   * Event types: `AUTO_MODERATION_BLOCK_MESSAGE`, `AUTO_MODERATION_FLAG_TO_CHANNEL`, `AUTO_MODERATION_USER_COMMUNICATION_DISABLED`
-   */
-  auto_moderation_rule_trigger_type?: string
-  /**
-   * Channel in which the entities were targeted.
-   *
-   * Event types: `MEMBER_MOVE`, `MESSAGE_PIN`, `MESSAGE_UNPIN`, `MESSAGE_DELETE`, `STAGE_INSTANCE_CREATE`, `STAGE_INSTANCE_UPDATE`, `STAGE_INSTANCE_DELETE`
-   */
-  channel_id?: string
-  /**
-   * Number of entities that were targeted.
-   *
-   * Event types: `MESSAGE_DELETE`, `MESSAGE_BULK_DELETE`, `MEMBER_DISCONNECT`, `MEMBER_MOVE`
-   */
-  count?: string
-  /**
-   * Number of days after which inactive members were kicked.
-   *
-   * Event types: `MEMBER_PRUNE`
-   */
-  delete_member_days?: string
-  /**
-   * ID of the overwritten entity.
-   *
-   * Event types: `CHANNEL_OVERWRITE_CREATE`, `CHANNEL_OVERWRITE_UPDATE`, `CHANNEL_OVERWRITE_DELETE`
-   */
-  id?: string
-  /**
-   * Number of members removed by the prune.
-   *
-   * Event types: `MEMBER_PRUNE`
-   */
-  members_removed?: string
-  /**
-   * ID of the message that was targeted.
-   *
-   * Event types: `MESSAGE_PIN`, `MESSAGE_UNPIN`, `STAGE_INSTANCE_CREATE`, `STAGE_INSTANCE_UPDATE`, `STAGE_INSTANCE_DELETE`
-   */
-  message_id?: string
-  /**
-   * Name of the role if type is "0" (not present if type is "1").
-   *
-   * Event types: `CHANNEL_OVERWRITE_CREATE`, `CHANNEL_OVERWRITE_UPDATE`, `CHANNEL_OVERWRITE_DELETE`
-   */
-  role_name?: string
-  /**
-   * Type of overwritten entity - "0", for "role", or "1" for "member".
-   *
-   * Event types: `CHANNEL_OVERWRITE_CREATE`, `CHANNEL_OVERWRITE_UPDATE`, `CHANNEL_OVERWRITE_DELETE`
-   */
-  type?: string
-  /**
-   * The type of integration which performed the action
-   *
-   * Event types: `MEMBER_KICK`, `MEMBER_ROLE_UPDATE`
-   */
-  integration_type?: string
 }
 
 export interface DiscordScheduledEvent {
