+++ packages/types/src/index.ts
@@ -1,5 +1,6 @@
 export * from './discord.js'
 export * from './discord/applications.js'
+export * from './discord/auditLogs.js'
 export * from './discord/gateway.js'
 export * from './discord/interactions.js'
 export * from './discord/reference.js'
