+++ packages/types/src/shared.ts
@@ -426,142 +426,6 @@ export enum StickerFormatTypes {
   Gif,
 }
 
-/** https://discord.com/developers/docs/resources/audit-log#audit-log-entry-object-audit-log-events */
-export enum AuditLogEvents {
-  /** Server settings were updated */
-  GuildUpdate = 1,
-  /** Channel was created */
-  ChannelCreate = 10,
-  /** Channel settings were updated */
-  ChannelUpdate,
-  /** Channel was deleted */
-  ChannelDelete,
-  /** Permission overwrite was added to a channel */
-  ChannelOverwriteCreate,
-  /** Permission overwrite was updated for a channel */
-  ChannelOverwriteUpdate,
-  /** Permission overwrite was deleted from a channel */
-  ChannelOverwriteDelete,
-  /** Member was removed from server */
-  MemberKick = 20,
-  /** Members were pruned from server */
-  MemberPrune,
-  /** Member was banned from server */
-  MemberBanAdd,
-  /** Server ban was lifted for a member */
-  MemberBanRemove,
-  /** Member was updated in server */
-  MemberUpdate,
-  /** Member was added or removed from a role */
-  MemberRoleUpdate,
-  /** Member was moved to a different voice channel */
-  MemberMove,
-  /** Member was disconnected from a voice channel */
-  MemberDisconnect,
-  /** Bot user was added to server */
-  BotAdd,
-  /** Role was created */
-  RoleCreate = 30,
-  /** Role was edited */
-  RoleUpdate,
-  /** Role was deleted */
-  RoleDelete,
-  /** Server invite was created */
-  InviteCreate = 40,
-  /** Server invite was updated */
-  InviteUpdate,
-  /** Server invite was deleted */
-  InviteDelete,
-  /** Webhook was created */
-  WebhookCreate = 50,
-  /** Webhook properties or channel were updated */
-  WebhookUpdate,
-  /** Webhook was deleted */
-  WebhookDelete,
-  /** Emoji was created */
-  EmojiCreate = 60,
-  /** Emoji name was updated */
-  EmojiUpdate,
-  /** Emoji was deleted */
-  EmojiDelete,
-  /** Single message was deleted */
-  MessageDelete = 72,
-  /** Multiple messages were deleted */
-  MessageBulkDelete,
-  /** Messaged was pinned to a channel */
-  MessagePin,
-  /** Message was unpinned from a channel */
-  MessageUnpin,
-  /** App was added to server */
-  IntegrationCreate = 80,
-  /** App was updated (as an example, its scopes were updated) */
-  IntegrationUpdate,
-  /** App was removed from server */
-  IntegrationDelete,
-  /** Stage instance was created (stage channel becomes live) */
-  StageInstanceCreate,
-  /** Stage instace details were updated */
-  StageInstanceUpdate,
-  /** Stage instance was deleted (stage channel no longer live) */
-  StageInstanceDelete,
-  /** Sticker was created */
-  StickerCreate = 90,
-  /** Sticker details were updated */
-  StickerUpdate,
-  /** Sticker was deleted */
-  StickerDelete,
-  /** Event was created */
-  GuildScheduledEventCreate = 100,
-  /** Event was updated */
-  GuildScheduledEventUpdate,
-  /** Event was cancelled */
-  GuildScheduledEventDelete,
-  /** Thread was created in a channel */
-  ThreadCreate = 110,
-  /** Thread was updated */
-  ThreadUpdate,
-  /** Thread was deleted */
-  ThreadDelete,
-  /** Permissions were updated for a command */
-  ApplicationCommandPermissionUpdate = 121,
-  /** Soundboard sound was created */
-  SoundboardSoundCreate = 130,
-  /** Soundboard sound was updated */
-  SoundboardSoundUpdate,
-  /** Soundboard sound was deleted */
-  SoundboardSoundDelete,
-  /** Auto moderation rule was created */
-  AutoModerationRuleCreate = 140,
-  /** Auto moderation rule was updated */
-  AutoModerationRuleUpdate,
-  /** Auto moderation rule was deleted */
-  AutoModerationRuleDelete,
-  /** Message was blocked by AutoMod according to a rule. */
-  AutoModerationBlockMessage,
-  /** Message was flagged by AutoMod */
-  AudoModerationFlagMessage,
-  /** Member was timed out by AutoMod */
-  AutoModerationMemberTimedOut,
-  /** Creator monetization request was created */
-  CreatorMonetizationRequestCreated = 150,
-  /** Creator monetization terms were accepted */
-  CreatorMonetizationTermsAccepted,
-  /** Guild Onboarding Question was created */
-  OnBoardingPromptCreate = 163,
-  /** Guild Onboarding Question was updated */
-  OnBoardingPromptUpdate,
-  /** Guild Onboarding Question was deleted */
-  OnBoardingPromptDelete,
-  /** Guild Onboarding was created */
-  OnBoardingCreate,
-  /** Guild Onboarding was updated */
-  OnBoardingUpdate,
-  /** Guild Server Guide was created */
-  HomeSettingsCreate = 190,
-  /** Guild Server Guide was updated */
-  HomeSettingsUpdate,
-}
-
 export enum ScheduledEventPrivacyLevel {
   /** the scheduled event is only accessible to guild members */
   GuildOnly = 2,
